Combined both csproj files

+++ PEReader/CorDirectory/Meta/Method/ElementType.cs
@@ -14,7 +14,7 @@ namespace AlphaOmega.Debug.CorDirectory.Meta
 		/// <summary>A single-dimensional, zero lower-bound array type modifier</summary>
 		public Boolean IsArray { get; internal set; }
 
-		/// <summary>As ingle-dimentional or multi-dimentional array type modifier</summary>
+		/// <summary>A single-dimensional or multi-dimensional array type modifier</summary>
 		public Byte MultiArray { get; internal set; }
 
 		/// <summary>A pointer type modifier</summary>
@@ -45,7 +45,7 @@ namespace AlphaOmega.Debug.CorDirectory.Meta
 			this.IsRequiredModifier = this.Type == Cor.ELEMENT_TYPE.CMOD_REQ;
 			if(this.IsRequiredModifier)
 			{
-				UInt32 rawValue1 = (UInt32)NativeMethods.GetPackedValue(signature, offset, out Int32 padding);
+				UInt32 rawValue1 = NativeMethods.GetPackedValue(signature, offset, out Int32 padding);
 				this.TypeDefOrRef = new CellPointerBase(cell, rawValue1);
 				this.RawPointer = rawValue1;
 				offset += (UInt32)padding;
@@ -79,14 +79,14 @@ namespace AlphaOmega.Debug.CorDirectory.Meta
 			case Cor.ELEMENT_TYPE.CLASS:// Followed by TypeDef or TypeRef token
 			case Cor.ELEMENT_TYPE.VALUETYPE:// Followed by TypeDef or TypeRef token
 			case Cor.ELEMENT_TYPE.CMOD_REQ:// Required modifier : followed by a TypeDef or TypeRef token
-				UInt32 rawValue1 = (UInt32)NativeMethods.GetPackedValue(signature, offset, out Int32 padding);
+				UInt32 rawValue1 = NativeMethods.GetPackedValue(signature, offset, out Int32 padding);
 				this.TypeDefOrRef = new CellPointerBase(cell, rawValue1);
 				this.RawPointer = rawValue1;
 				offset += (UInt32)padding;
 				break;
 			case Cor.ELEMENT_TYPE.MVAR:
 			case Cor.ELEMENT_TYPE.VAR:
-				UInt32 genericArgumentIndex = (UInt32)NativeMethods.GetPackedValue(signature, offset, out Int32 padding2);
+				UInt32 genericArgumentIndex = NativeMethods.GetPackedValue(signature, offset, out Int32 padding2);
 				this.TypeDefOrRef = null;
 				this.RawPointer = genericArgumentIndex;
 				offset += (UInt32)padding2;
@@ -103,7 +103,7 @@ namespace AlphaOmega.Debug.CorDirectory.Meta
 				this.GenericArguments = null;
 		}
 
-		/// <summary>Formats element as it storead in CLI</summary>
+		/// <summary>Formats element as it stored in CLI</summary>
 		/// <returns>String representation</returns>
 		public String ToStringOriginal()
 		{
